[update] move search screen to /screens to use
